Refactoring class
